[cuda] fix int32 overflow in quantized histogram kernels' data indexing
CUDAConstructDiscretizedHistogramDenseKernel (shared and global memory
variants) computed the partition base offset
(partition_column_start * num_data) and the per-row offset
(data_index * num_columns_in_partition) in 32-bit arithmetic. On wide
datasets the products exceed 2^31 and wrap: reads land out of bounds
(illegal memory access, observed on 5.4M x 2748 numerai data) or,
below the crash threshold, inside the buffer -- silently accumulating
histograms from wrong bins. The non-quantized kernels already widen
these products to size_t; the quantized kernels were missed.

Verified on numerai v5 (2748 features):
- pre-fix: crash at >=2.09M rows; at 1.65M rows runs but predictions
  correlate only 0.812 with the CPU quantized reference
- post-fix: 5.43M rows trains cleanly; predictions match CPU quantized
  reference with corr 1.000000 (max abs diff 8e-8, 30 trees)
- compute-sanitizer memcheck: invalid 1-byte global reads at
  cuda_histogram_constructor.cu:574 gone

+++ src/treelearner/cuda/cuda_histogram_constructor.cu
@@ -546,7 +546,7 @@ __global__ void CUDAConstructDiscretizedHistogramDenseKernel(
   const unsigned int num_threads_per_block = blockDim.x * blockDim.y;
   const int partition_column_start = feature_partition_column_index_offsets[blockIdx.x];
   const int partition_column_end = feature_partition_column_index_offsets[blockIdx.x + 1];
-  const BIN_TYPE* data_ptr = data + partition_column_start * num_data;
+  const BIN_TYPE* data_ptr = data + static_cast<size_t>(partition_column_start) * num_data;
   const int num_columns_in_partition = partition_column_end - partition_column_start;
   const uint32_t partition_hist_start = column_hist_offsets_full[blockIdx.x];
   const uint32_t partition_hist_end = column_hist_offsets_full[blockIdx.x + 1];
@@ -571,7 +571,7 @@ __global__ void CUDAConstructDiscretizedHistogramDenseKernel(
     for (data_size_t i = 0; i < num_iteration_this; ++i) {
       const data_size_t data_index = data_indices_ref_this_block[inner_data_index];
       const int32_t grad_and_hess = cuda_gradients_and_hessians[data_index];
-      const uint32_t bin = static_cast<uint32_t>(data_ptr[data_index * num_columns_in_partition + threadIdx.x]);
+      const uint32_t bin = static_cast<uint32_t>(data_ptr[static_cast<size_t>(data_index) * num_columns_in_partition + threadIdx.x]);
       int32_t* pos_ptr = shared_hist_ptr + bin;
       atomicAdd_block(pos_ptr, grad_and_hess);
       inner_data_index += blockDim.y;
@@ -676,7 +676,7 @@ __global__ void CUDAConstructDiscretizedHistogramDenseKernel_GlobalMemory(
   const unsigned int num_threads_per_block = blockDim.x * blockDim.y;
   const int partition_column_start = feature_partition_column_index_offsets[blockIdx.x];
   const int partition_column_end = feature_partition_column_index_offsets[blockIdx.x + 1];
-  const BIN_TYPE* data_ptr = data + partition_column_start * num_data;
+  const BIN_TYPE* data_ptr = data + static_cast<size_t>(partition_column_start) * num_data;
   const int num_columns_in_partition = partition_column_end - partition_column_start;
   const uint32_t partition_hist_start = column_hist_offsets_full[blockIdx.x];
   const uint32_t partition_hist_end = column_hist_offsets_full[blockIdx.x + 1];
@@ -703,7 +703,7 @@ __global__ void CUDAConstructDiscretizedHistogramDenseKernel_GlobalMemory(
     for (data_size_t i = 0; i < num_iteration_this; ++i) {
       const data_size_t data_index = data_indices_ref_this_block[inner_data_index];
       const int32_t grad_and_hess = cuda_gradients_and_hessians[data_index];
-      const uint32_t bin = static_cast<uint32_t>(data_ptr[data_index * num_columns_in_partition + threadIdx.x]);
+      const uint32_t bin = static_cast<uint32_t>(data_ptr[static_cast<size_t>(data_index) * num_columns_in_partition + threadIdx.x]);
       int32_t* pos_ptr = shared_hist_ptr + bin;
       atomicAdd_block(pos_ptr, grad_and_hess);
       inner_data_index += blockDim.y;
